backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# Settings
pydantic-settings==2.7.1

# HTTP client
httpx==0.28.1

# Document Processing
PyMuPDF==1.25.3
pytesseract==0.3.13
Pillow==11.1.0

# Embeddings & Vector Store
qdrant-client==1.12.1
tiktoken==0.8.0

# LLM Clients
groq==0.15.0
google-generativeai==0.8.4
openai==1.59.6

# Async Tasks
celery[redis]==5.4.0

# Utilities
python-dotenv==1.0.1
setuptools

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# ── OPTIONAL: Video/Audio transcription ──────────────────
# Install these separately if you need video/audio processing:
#   pip install openai-whisper ffmpeg-python
# Whisper requires PyTorch — see https://pytorch.org/get-started
# ffmpeg-python==0.2.0
# openai-whisper==20240930
# pip install openai-whisper ffmpeg-python